one-road step 0: rescue path helpers to scan/package-paths.ts
